facial_recognition/requirements.txt
numpy
opencv-python
Pillow
retinaface
